requirements.txt
requests
PyPDF2
pdf2image
pytesseract
Pillow
openai>=1.0
supabase>=2.0
python-dotenv
tqdm
pandas
